+++ src/main/java/com/bankapi/bank/model/AccountDTO.java
@@ -1,17 +1,15 @@
 package com.bankapi.bank.model;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 @Data
 public class AccountDTO {
-    private BigDecimal creditLimit;
+    private double creditLimit;
 
-    private BigDecimal balance;
+    private Double balance;
     
-    private BigDecimal debt;
+    private Double debt;
     
     @NotNull
     private Long clientId;
